requirements.txt
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
coverage==7.6.1
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
Werkzeug==3.0.4
